+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 - Prestige formula changed to 100x[RATIO] - 40 + [LAND]/150 (from 120x[RATIO] - 50 + [LAND]/250)
 - Tech cost changed to 2.5x[LAND] + 100x[TECHS] (from 2.5x[LAND] + 150x[TECHS])
 - Wizard resilience is again gained immediately, 10 per successful fireball or lightning bolt
-- Wizard resilience now increases peasant change by [RESIL/200]% of your population deficit
+- Wizard resilience now increases peasant change by [RESIL/125]% of your population deficit
 - Fireball damage increased to 2.75% (from 2.5%)
 - Lightning bolt repair invest bonus replaced with temporary damage portion based on wizard resilience, [RESIL/40]% of damage is automatically repaired in 12 hours
 - Black op spells with duration changed to 8 hours, +2 in war, +4 in mutual war (from 6, +3, +6)